+++ CHANGELOG.md
@@ -11,6 +11,7 @@ This file contains the list of changes made to pyjoulescope_ui.
 *   Improved device open error handling
 *   Managed future features
 *   Added waveform options: hide min/max signals, show grid, trace width.
+*   Display Joulescope driver version in ABOUT.
 
 
 ## 0.2.6